move the logging functions into a separate file
